requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
inquirer==3.1.3
